requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
autograd==1.8.0
autoray==0.8.2
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
cachetools==7.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diastatic-malt==2.15.2
executing==2.2.1
fastapi==0.104.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.10
iniconfig==2.3.0
intel-openmp==2021.4.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mistune==3.1.4
mkl==2021.4.0
mpmath==1.3.0
multidict==6.7.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.0.2
notebook_shim==0.2.4
numpy==1.24.4
overrides==7.7.0
packaging==25.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.1
pennylane==0.42.0
pennylane_lightning==0.44.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
psycopg2-binary==2.9.7
PuLP==3.3.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.0.9
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
qtconsole==5.7.0
QtPy==2.4.3
rdkit==2025.9.4
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rustworkx==0.17.1
safetensors==0.6.2
scikit-learn==1.3.0
scipy==1.11.2
scipy-openblas32==0.3.31.22.1
seaborn==0.12.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.20
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.5
sympy==1.13.1
tbb==2021.13.1
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.21
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.14.0
torch==2.3.1+cpu
torch-geometric==2.7.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.22.0
